[IMP]stock_available_immediately_exclude_location: Improve perfs
* Replace specific code to compute the location domain from the stock quant domain by a call to a method provided by the stock_available_location_get_domain addon. This new method significantly improves the performances by avoiding the call to the 'fields_get' method on the model. This reduce the processing time to compute the domain from 17.4ms to less than 1ms.
* Caching the list of locations to exclude from the immediately available
  quantity computation.
* Avoiding to query the incoming and outgoing quantities for the locations to
  exclude when computing the available quantity into the locations to exclude.

As result, on the workbench used when profiling the module, the time to compute
the available quantity is reduced from 158ms to 134ms (15% improvement). If we
take into account the performance gain of the previous commit, the time is
reduced from 244ms to 134ms (45% improvement).

* Restore the original behaviour on computing the available quantity
for excluded locations on strict mode. IOW the available quantity is computed
on records that are linked directly to the location_id and not on records that
are linked to a location child of the excluded locations.

This is a performance improvement and boost the computation from 134ms to 74ms (45% improvement) on the workbench test.