[IMP]stock_available_immediately_exclude_location: Improve perfs and remove duplicate code
Replace specific code to compute the location domain from the stock quant domain by a call to a method provided by the stock_available_location_get_domain addon. This new method significantly improves the performances by avoiding the call to the 'fields_get' method on the model. This reduce the processing time to compute the domain from 17.4ms to less than 1ms.