[IMP]stock_available_immediately_exclude_location: Improve perfs
This commit is a follow-up of the previous one. It improves the perfs of the
module by:

* Caching the list of locations to exclude from the immediately available
  quantity computation.
* Avoiding to query the incoming and outgoing quantities for the locations to
  exclude when computing the available quantity into the locations to exclude.

As result, on the workbench used when profiling the module, the time to compute
the available quantity is reduced from 158ms to 134ms (15% improvement). If we
take into account the performance gain of the previous commit, the time is
reduced from 244ms to 134ms (45% improvement).