[IMP] stock_available_immediately_exclude_location: Improve perfs
This commit restore the original behaviour on computing the available quantity
for excluded locations on strict mode. IOW the available quantity is computed
on records that are linked directly to the location_id and not on records that
are linked to a location child of the excluded locations. This is a performance
improvement and boost the computation from 134ms to 74ms (45% improvement)
on the workbench test.